OpenDSSC: Fix pointers and function signatures for 32-bit Windows
git-svn-id: https://svn.code.sf.net/p/electricdss/code/trunk@3793 d8739450-1e93-4ef4-a0af-c327d92816ff

+++ VersionC/DLL/OpenDSSCDLL.cpp
@@ -524,7 +524,7 @@ void __stdcall DSSV(int mode, uintptr_t* myPtr, int* myType, int* mySize)
 		WriteStr2Array("Error, parameter not recognized");
 		WriteStr2Array(Char0());
 		*mySize = myStrArray.size();
-		myPtr = (uint64_t*)&(myStrArray[0]);
+		myPtr = (uintptr_t*)&(myStrArray[0]);
 	}
